+++ packages/apis/example/lib/services/handlers/discount_handlers/price_rule_handler/create_price_rule_free_item_handler.dart
@@ -1,5 +1,5 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/discounts/price_rule/abstract/price_rule.dart';
+import 'package:apis/network/remote/discounts/price_rule/abstract/price_rule_service.dart';
 import 'package:apis/network/remote/discounts/price_rule/freezed_model/request/create_price_rule_free_item_request.dart'
     as create_request;
 import 'package:apis/network/remote/discounts/price_rule/freezed_model/response/create_price_rule_free_item_response.dart';
