+++ packages/apis/example/lib/services/handlers/discount_handlers/price_rule_handler/retrieves_count_of_price_rules_handler.dart
@@ -1,4 +1,4 @@
-import 'package:apis/network/remote/discounts/price_rule/abstract/price_rule.dart';
+import 'package:apis/network/remote/discounts/price_rule/abstract/price_rule_service.dart';
 import 'package:apis/network/remote/discounts/price_rule/freezed_model/response/price_rule_list_response.dart';
 import 'package:apis/apis.dart';
 import 'package:example/services/api_request_handler.dart';
